feat(components): Refactor/Clean Up Combobox Content (#1550)
* extract search and list into own components

* use classnames conditional syntax

* reorganize hook content to colocate related logic

* use ref to avoid one more dependency on listener useEffect

* extract hooks to own files and directory

* add unit tests for search

* add unit tests for list, convert some tests into unit tests instead

* merge feature branch, resolve conflicts, move styles

* move div out of ul, fix styles on fade and peek height
