+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContent.css
@@ -19,96 +19,11 @@
   visibility: hidden;
 }
 
-.clearSearch {
-  position: absolute;
-  top: 50%;
-  right: var(--space-base);
-  z-index: var(--elevation-tooltip);
-  width: var(--space-large);
-  height: var(--space-large);
-  padding: 0;
-  border: none;
-  border-radius: var(--radius-circle);
-  background-color: var(--color-surface--background);
-  cursor: pointer;
-  transform: translateY(-50%);
-}
-
-.optionsList {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-smaller);
-  max-height: calc(var(--space-small) * 32.5);
-  margin: 0;
-  padding: var(--space-small);
-  overflow: scroll;
-  list-style: none;
-}
-
-.option {
-  display: flex;
-  min-height: calc((var(--space-largest) - var(--space-smaller)));
-  box-sizing: border-box;
-  padding: var(--space-small);
-  border-radius: var(--radius-large);
-  color: var(--color-heading);
-  font-weight: 500;
-  cursor: pointer;
-  justify-content: space-between;
-  align-items: center;
-  transition: all var(--timing-base);
-}
-
-.selectedOption,
-.option:hover,
-.option:focus {
-  background-color: var(--color-surface--background);
-}
-
-.option:focus,
-.clearSearch:focus {
-  box-shadow: var(--shadow-focus);
-}
-
-.search {
-  position: relative;
-}
-
-.searchInput {
-  width: 100%;
-  padding: var(--space-base);
-  padding-right: calc(var(--space-large) * 2.25);
-  border: none;
-  border-bottom: var(--border-base) solid var(--color-border);
-  font-family: var(--typography--fontFamily-normal);
-  font-size: var(--typography--fontSize-base);
-}
-
-.filterMessage,
-.emptyStateMessage {
-  padding: var(--space-small);
-}
-
-.search::after {
-  content: "";
-  display: block;
-  position: absolute;
-  right: 0;
-  bottom: calc(var(--space-base) * -1);
-  left: 0;
-  height: var(--space-base);
-  background-image: linear-gradient(
-    to bottom,
-    rgba(255, 255, 255, 1),
-    rgba(255, 255, 255, 0)
-  );
-}
-
 .actions::before {
   content: "";
   display: block;
   position: absolute;
-  top: calc(var(--space-base) * -1);
+  top: calc(var(--space-large) * -1);
   right: 0;
   left: 0;
   height: var(--space-base);
