+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContent.css.d.ts
@@ -1,14 +1,6 @@
 declare const styles: {
   readonly "content": string;
   readonly "hidden": string;
-  readonly "clearSearch": string;
-  readonly "optionsList": string;
-  readonly "option": string;
-  readonly "selectedOption": string;
-  readonly "search": string;
-  readonly "searchInput": string;
-  readonly "filterMessage": string;
-  readonly "emptyStateMessage": string;
   readonly "actions": string;
   readonly "actionPadding": string;
 };
