+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxList.css
@@ -1,0 +1,43 @@
+.container {
+  padding: var(--space-small);
+}
+
+.optionsList {
+  display: flex;
+  max-height: calc(var(--space-small) * 33.33);
+  margin: 0;
+  padding: 0;
+  overflow: scroll;
+  list-style: none;
+  flex-direction: column;
+  gap: var(--space-smaller);
+}
+
+.option {
+  display: flex;
+  min-height: calc((var(--space-largest) - var(--space-smaller)));
+  box-sizing: border-box;
+  padding: var(--space-small);
+  border-radius: var(--radius-large);
+  color: var(--color-heading);
+  font-weight: 500;
+  cursor: pointer;
+  justify-content: space-between;
+  align-items: center;
+  transition: all var(--timing-base);
+}
+
+.selectedOption,
+.option:hover,
+.option:focus {
+  background-color: var(--color-surface--background);
+}
+
+.option:focus {
+  box-shadow: var(--shadow-focus);
+}
+
+.filterMessage,
+.emptyStateMessage {
+  padding: var(--space-small);
+}
