+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxList.css.d.ts
@@ -1,0 +1,10 @@
+declare const styles: {
+  readonly "container": string;
+  readonly "optionsList": string;
+  readonly "option": string;
+  readonly "selectedOption": string;
+  readonly "filterMessage": string;
+  readonly "emptyStateMessage": string;
+};
+export = styles;
+
